[apps/shared] Resolve conflict: function tmin and tmax are float instead of double

+++ apps/shared/function.h
@@ -36,8 +36,8 @@ public:
 
   // Definition Interval
   virtual bool shouldClipTRangeToXRange() const { return true; } // Returns true if the function will not be displayed if t is outside x range.
-  virtual double tMin() const { return NAN; }
-  virtual double tMax() const { return NAN; }
+  virtual float tMin() const { return NAN; }
+  virtual float tMax() const { return NAN; }
   virtual float rangeStep() const { return NAN; }
 
   // Name
